requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
djangorestframework==3.11.0
idna==2.9
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
unittest-xml-reporting==3.0.2
urllib3==1.25.9
asgiref==3.2.7
attrs==21.4.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
djangorestframework==3.11.0
idna==2.9
iniconfig==1.1.1
jsonfield==3.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
tomli==2.0.1
unittest-xml-reporting==3.0.2
urllib3==1.25.9
